Connect MCU_RDY to ~READY pin
In the old designs, the ~READY pin was never used outside of the stack
and the ~MCU_RDY pin was used in its place. I want to now make it so
that using ~READY will always be synchronized. But the older rev1 cards
directly connect ready to the i960Sx